backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
django-filter==24.3
django-environ==0.11.2
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-redis==5.4.0
django-extensions==3.2.3
django-import-export==4.3.4

psycopg2-binary==2.9.10
redis==5.2.1
celery==5.4.0

gunicorn==23.0.0
whitenoise==6.8.2

Pillow==11.1.0
openpyxl==3.1.5
reportlab==4.2.5
python-dateutil==2.9.0

sentry-sdk==2.19.2
drf-spectacular==0.28.0
